add tracker scripts
This is taken from plausible. A simple find and replace of plausible with vince is done.
It does not matter the name of the script, but we are avoiding  plausible branding
since we are competitor.

The plan is to embed the generated js files and serve them from virtual file system.
This change does not include embedding.