initial timeseries.RootQuery implementation
This allows querying aggregates for specific [property, metric, key ]
combination. Base property behave differently, for Base we query all properties
for selected metric while selecting all metrics for Base property.

** fixed value versions **

We rely on versions for querying. This commit uses managed badger, where
we manually set commit timestamp which is rounded to Milliseconds . This
timestamp is used when querying, as it represents time when the aggregate
was computed.

Previously we were using un managed badger, which was operating under wrong
assumptions. This bug wouldn't have happened if we had unit tests to validate
this functionality.