use slice for db shards
using a map is a an overkill, a slice with binary search is much better
since we are guaranteed to have always smaller number of active shards.

For now we load all shards in memory, for optimal utilization older shards must
be released and only open when needed. We already have a simple lru cache which
fits this use case.

We can revisit this implementation when needs changes.