sublime like license
Limit trial to up to 600 page views per month and unlimited for licensed
instances.

Treat each instance as a single unit. This simplify license management, the
model is inspired by sublime text with some notable modifications.

-
The trial period limits up to 600 events per month. The rate limiter never
resets after this but users are free to restart the instance to reset the timer.

-
The admin account must match the email address set in the license key. The license
key lifts all restictions from using the vince instance.

-
The lincese keys are valid for 2 years instead of 3 like sublime. vince is a server
and a lot of efforts goes to mintaining it. 3 years is too long to trigger renewals
and the pricing will not be attractive for small sites.

Like sublime we only watch the release builds. vince versions now have a date
as part of the semver eg `v0.0.0+20240824` as long as the lincense key expiry date
is before the build then the user is free to use the upgrade and get support.