save one event at a time
We already use channel as a queue, no need for secondary buffer. The reason
we are doing this is most cases we spend out time waiting for events.

To optimize for keeping lots of events instances in memory. We will be writing
them as sooon as tehy leave the cache.