reuse bitmap in tx.Row
we use a bitmap to compute  key range. Previously we were discarding it
after determining the key range.

This commit reuses the same bitmap to return the result avoiding extra
allocations.

We also add benchmark for future reference because tx.Row is the heart
of search, any gains here have a bigger impact on query performance.