avoid blind views scan
Peviously we were hitting the key value store for all generated views from
interval. This commit keeps a  bitmap of all views observed in a shard. This
ensures that we only iterate for views actually registered in the database.

We save a lot of key/value lookups for generated views not seen in the
database.