perf: use file based buffers for translation data
Translation takes a huge chunk of allocations during ingestion. We rely
heavily on translation trie and btree for search.

Our primary objective is to have decent performance on cheap cloud instances
which roughly has 1 vCpu and 2 GB of ram and 24 GB of storage.

Using mmap files for buffers works better in this scenario, we leave  the
rest of ram towards key/value store scans and buffers.